Melhorias no backup de servidores

+++ src/main/java/com/kuuhaku/command/commands/information/PingCommand.java
@@ -46,7 +46,7 @@ public class PingCommand extends Command {
 	@Override
 	public void execute(User author, Member member, String rawCmd, String[] args, Message message, MessageChannel channel, Guild guild, String prefix) {
 		int fp = (int) ((Runtime.getRuntime().totalMemory() - Runtime.getRuntime().freeMemory()) / 1024 / 1024);
-		int ping = Math.toIntExact(Main.getInfo().getAPI().getRestPing().complete());
+		long ping = Main.getInfo().getAPI().getRestPing().complete();
 		channel.sendMessage(":ping_pong: Pong! ")
 				.flatMap(m -> m.editMessage(m.getContentRaw() + " " + ping + " ms!"))
 				.flatMap(m -> m.editMessage(m.getContentRaw() + "\n:floppy_disk: " + fp + " MB!"))
